Fix a bug with projectile.Firer, add teleporter
THIS TOOK SO LONG TO TRACK DOWN luckily I was in a boring lecture with a
friend who first encountered the effects of this bug!

Fixed the Firer component of a projectile being a copy of the host
entity instead of a reference to the actual entity.

Also fixed a bug where the projectile spawning function returned the
template of the projectile instead of the actual entity.

This all because of trying to add a teleporter projectile. Added that
too.